upi momo jeki kirana sudah di mergee, Jesper tinggal di merge, tompel belum di merge

+++ database/migrations/2019_03_18_141451_create_kemajuan_proyeks_table.php
@@ -29,7 +29,7 @@ class CreateKemajuanProyeksTable extends Migration
             $table->integer('tipeKemajuan');
 
 
-            $table->integer('value');
+            $table->bigInteger('value');
 
             $table->bigInteger('pelaksanaan_id')->unsigned();
             $table->foreign('pelaksanaan_id')
